+++ src/a2x_pack_console.c
@@ -149,8 +149,8 @@ void a_console__init2(void)
     a_sprite_free(graphics);
 
     a_font_setFace(A_FONT_FACE_WHITE);
-    g_linesPerScreen = a_settings_getInt("video.height")
-                            / a_font_getLineHeight() - 2;
+    g_linesPerScreen = (unsigned)(a_settings_getInt("video.height")
+                                  / a_font_getLineHeight() - 2);
 
     // In case messages were logged between init and init2
     while(a_list_size(g_lines) > g_linesPerScreen) {
